Stabilize the current `#[builder(getter)]` feature MVP (#251)
It's been ~3 months since we've introduced an MVP of the
`#[builder(getter)]` attribute, and the feature hasn't changed since its
initial design. There are already some people using this API and asking
to get it stabilized.

I believe the current API design is already firm, and still amenable to
the future extensions described in
https://github.com/elastio/bon/issues/225, so let's deploy this 🚀