Trigger a compiler warning if `#[tracing::instrument]` is placed after the `#[builder]` instead of before (#264)
Closes https://github.com/elastio/bon/issues/259.

A new warning suggests moving the `#[tracing::instrument]` before the
`#[builder]` macro. Doing it as a warning because this check is
heuristic and may technically have false positives for macros that use
the name `instrument` too. I don't know of any such macros, and if there
are ones, I expect people will reach out and we'll work on a better
solution.