Some improvements to the doc test (see commit body below for details)
Use `#[tokio::main]` instead of `tokio_test` in the doc example to make it a more real-world-looking.

Also strip the `std::io::Result` for simplicity. Async functions are usually fallible, but for the purposes of an example, it's fine to simplify it. Besides, to my personal mind `std::io::Error`  is usually a bad pattern as it returns a non-descriptive error message. I'd promote `anyhow` instead in the docs, but here I think making it infallible is okay.